Move Carousel.jsx to components folder and update links accordingly
